+++ TP3/src/main/java/library/model/IDAO.java
@@ -3,7 +3,7 @@ package main.java.library.model;
 import java.util.List;
 
 public interface IDAO<T> {
-  T findById(Long id);
+  T findById(int id);
 
   List<T> getAll();
 
@@ -11,5 +11,5 @@ public interface IDAO<T> {
 
   boolean update(T entity);
 
-  boolean delete(Long id);
+  boolean delete(int id);
 }
